created new classes
for more comfortable using classes
