+++ hrmsAPI/backend/bookings.py
@@ -1,0 +1,14 @@
+from sqlalchemy import create_engine, Column, Integer, String, Boolean, DateTime, Text, ForeignKey
+from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.orm import sessionmaker, relationship
+
+Base = declarative_base()
+
+class Bookings(Base):
+    __tablename__ = 'bookings'
+    id = Column(Integer, primary_key=True)
+    user_id = Column(Integer, ForeignKey('Users.id'))
+    table_id = Column(Integer, ForeignKey('Tables.id'))
+    persons = Column(Integer)
+    date = Column(DateTime)
+    comment = Column(Text)
