+++ hrmsAPI/backend/users.py
@@ -1,0 +1,13 @@
+from sqlalchemy import create_engine, Column, Integer, String, Boolean, DateTime, Text, ForeignKey
+from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.orm import sessionmaker, relationship
+
+Base = declarative_base()
+
+class Users(Base):
+    __tablename__ = 'users'
+    id = Column(Integer, primary_key=True)
+    name = Column(String(100))
+
+    orders = relationship('Orders', backref='user')
+    bookings = relationship('Bookings', backref='user')
